fix: YGText 삭제 후 YGAlert, YGLabel로 구분 및 padding 공통 적용
